Add remaining assignment files via upload
Added files via upload from my OneDrive. This includes the files with code and the relevant prompts and deliverables. Some files were too large to include in the upload. I kept the zipped files that were smaller so that you may extract them. You will need a WINRAR application if on Windows or 7-ZIP to unzip the rar files. I did not include the 2691 rar file for Assignment 4 as it was too large and not needed for the assignment anyways. Follow the prompt instructions if you do need it and download it the same way.